Predict the win possibilities of deals/projects for an IT company
We will analyze and predict the win possibilities of deals/projects for an IT consulting company and see how the possibility of winning a 
deal is impacted by other variables. This will enable the IT consulting company to manage the effort required to win a deal to meet the growth 
targets.